% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{BWInteracted}
\alias{BWInteracted}
\title{Check If All Between-Subject Factors are Interacted with Within-Subject Factors}
\usage{
BWInteracted(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}
}
\value{
This function will return a logical (\code{TRUE}/\code{FALSE}).
}
\description{
This function determines whether all between-subject terms are interacted
with the within-subject terms.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
BWInteracted(glsMod)
glsMod <- gls(Distance ~ Sex + Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
BWInteracted(glsMod)

}
\seealso{
{\code{\link[mmemmuris]{isReversible}}, \code{\link[mmemmuris]{completeData}}}
}
