% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{completeData}
\alias{completeData}
\title{Complete Cases}
\usage{
completeData(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm}.}
}
\value{
\tabular{ll}{
\code{complete}  \tab A logical telling the user if there are no missing values in the dataset. \cr
\code{N}  \tab The total number of observations of data in the long or wide format. \cr
\code{r}  \tab The total number of repeated measurements for data in the long format. \cr
\code{n}  \tab The total number of observations of data used in the long or wide format. \cr
\tab \cr
}
}
\description{
This function lets you know if you have any missing data in your dataset.
}
\details{
As introduced in Wright (1994) and (n.d.), the number of observations for
marginal and mixed-effects models is set equal to \deqn{n = N / r} where
\eqn{N} is the total number of observations of data in the long format and
\eqn{r} is the number of repeated measurements.  For multivariate linear models,
\eqn{N} is the total number of observations of data in the wide format and
\eqn{n} is the total number of complete observations of data in the wide format.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
completeData(glsMod)

}
\references{
{Wright, S. P. (1994). Adjusted F Tests for Repeated Measures with
the MIXED Procedure. Knoxville: Statistics Department, University of Tennessee.
\url{https://support.sas.com/resources/papers/proceedings-archive/SUGI95/Sugi-95-195\%20Wright.pdf}}

{Wright, S.P. (n.d.). Multivariate Analysis Using the MIXED Procedure. Sugi
23: Multivariate analysis using the mixed procedure - SAS support.
\url{https://support.sas.com/resources/papers/proceedings/proceedings/sugi23/Stats/p229.pdf}}
}
\seealso{
{\code{\link[mmemmuris]{isReversible}}, \code{\link[mmemmuris]{BWInteracted}}}
}
