% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{covStruct}
\alias{covStruct}
\title{Covariance Structure}
\usage{
covStruct(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}
}
\value{
This function will return a character saying which covariance structure the
model has:  \tabular{lll}{\cr
\code{un} \tab - \tab unstructured \cr
\code{cs} \tab - \tab compound symmetry \cr
\code{other} \tab - \tab not unstructured or compound symmetry \cr
}
}
\description{
This function will determine whether a compound symmetry, unstructured, or
other covariance structure was used for the random/repeated effects for
objects of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.
}
\details{
This package uses two different covariance structures: compound symmetry
(repeated measures ANOVA) and unstructured (multivariate analysis of
variance).  The compound symmetry covariance structure uses two (co)variance
parameters: a residual error variance (\eqn{\sigma ^ 2}) and covariance
parameter between repeated measurements (\eqn{\sigma_r ^ 2}).  The
unstructured covariance structure has the maximum number possible of
(co)variance parameters: \deqn{r(r + 1) / 2}
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
covStruct(glsMod)
# Compound symmetry covariance
glsModCS <- gls(Distance ~ Sex * Age,
                correlation = corCompSymm(form = ~ 1 | Subject),
                na.action = na.omit,
                data = orthodontLong)
covStruct(glsModCS)

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
covStruct(lmeMod)
# Compound symmetry covariance (random intercepts)
lmeModCS <- lme(Distance ~ Sex * Age,
                random = ~ 1 | Subject,
                na.action = na.omit,
                data = orthodontLong)
covStruct(lmeModCS)

}
\references{
{\url{https://support.sas.com/resources/papers/proceedings/proceedings/sugi30/198-30.pdf}}
}
