% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{ddfBW}
\alias{ddfBW}
\title{Between-Within Denominator Degrees of Freedom}
\usage{
ddfBW(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}
}
\value{
\tabular{ll}{
\code{between}  \tab A character vector containing the between-subject effects
for the model. \cr
\code{within}  \tab A character vector containing the within-subject effects
for the model. \cr
\code{bDenDF}  \tab A vector containing the between-subject degrees of freedom. \cr
\code{wDenDF}  \tab A vector containing the within-subject degrees of freedom. \cr
\code{rDenDF}  \tab A vector containing the residual degrees of freedom. \cr
\verb{Den df}  \tab A vector containing the denominator degrees of freedom
for each term in the model. \cr
\code{Xbetween}  \tab The between-subject effects design matrix. \cr
\tab \cr
}
}
\description{
This function calculates the between-within denominator degrees of freedom for
marginal (\code{\link[nlme]{gls}}) and mixed-effects (\code{\link[nlme]{lme}}) models.
}
\details{
The residual denominator degrees of freedom are calculated as

\eqn{ddf = N - rank(}\strong{X}\eqn{)}

where \eqn{N} is the total number of observations of data in the long format
and \strong{X} is the design matrix of the independent variables.

The denominator degrees of freedom are equal to

\eqn{ddfBetween = n - rank(}\strong{X_b} \eqn{)}

where \eqn{n} is the total number of observations of data in the wide format
and \strong{X_b} is the design matrix of the between-subject independent
variables in the wide format

for between-subject effects and

\eqn{ddfWithin = ddf - ddfBetween}

for within-subject effects.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
ddfBW(glsMod)
# Compound symmetry covariance
glsModCS <- gls(Distance ~ Sex * Age,
                correlation = corCompSymm(form = ~ 1 | Subject),
                na.action = na.omit,
                data = orthodontLong)
ddfBW(glsModCS)

}
\references{
{\url{https://documentation.sas.com/doc/en/statug/15.2/statug_glimmix_details37.htm}}
}
\seealso{
{\code{\link[mmemmuris]{ddfResidual}}}
}
