% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{ddfResidual}
\alias{ddfResidual}
\title{Residual Denominator Degrees of Freedom}
\usage{
ddfResidual(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}
}
\value{
\tabular{ll}{
\code{between}  \tab A character vector containing the between-subject effects
for the model. \cr
\code{within}  \tab A character vector containing the within-subject effects
for the model. \cr
\verb{Den df}  \tab A vector containing the denominator degrees of freedom
for each term in the model. \cr
\code{Xbetween}  \tab The between-subject effects design matrix. \cr
\code{X}  \tab The fixed effects design matrix. \cr
\tab \cr
}
}
\description{
This function calculates the residual denominator degrees of freedom for marginal
(\code{\link[nlme]{gls}}) and mixed-effects (\code{\link[nlme]{lme}}) models.
}
\details{
The residual denominator degrees of freedom are calculated as

\eqn{ddf = N - rank(}\strong{X}\eqn{)}

where \eqn{N} is the total number of observations of data in the long format
and \strong{X} is the design matrix of the independent variables.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
ddfResidual(glsMod)
# Compound symmetry covariance
glsModCS <- gls(Distance ~ Sex * Age,
                correlation = corCompSymm(form = ~ 1 | Subject),
                na.action = na.omit,
                data = orthodontLong)
ddfResidual(glsModCS)
}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.3/statug/statug_glimmix_details38.htm}}
}
