% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{Ematrix}
\alias{Ematrix}
\title{MANOVA/RM ANOVA SSCP Error (\strong{E}) Matrix}
\usage{
Ematrix(fMod, individual = NULL, ss = NULL)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm}.}

\item{individual}{The subject whose marginal variance-covariance matrix
should be used, where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design
matrix, \strong{G} is the corresponding (co)variances of the random effects matrix,
and \strong{R} represents the (co)variances of the repeated effects matrix).  The
default is the first subject in the factor.}
}
\value{
\tabular{ll}{
\code{E} \tab A sum of squares cross products (SSCP) error matrix for
within-subject effects. \cr
\tab \cr
\code{M} \tab A sum-to-zero contrast matrix for within-subject effects. \cr
\tab \cr
\code{SSE} \tab The error sum of squares for within-subject effects calculated
as \eqn{SS_E = trace(}\strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1})}. \cr
\tab \cr
\code{EB} \tab A sum of squares cross products (SSCP) error matrix for
between-subject effects. \cr
\tab \cr
\code{MB} \tab A sum-to-zero contrast matrix for between-subject effects. \cr
\tab \cr
\code{Vmatrix} \tab The \strong{V} matrix from the marginal or mixed-effects model. \cr
\tab \cr
\code{X} \tab A between-subject effects design matrix. \cr
\tab \cr
\code{n} \tab The sample size. \cr
\tab \cr
\code{r}  \tab The total number of repeated measurements for data in the long format. \cr
\tab \cr
}
}
\description{
This function returns the error (\strong{E}) sums of squares and crossproducts (SSCP)
matrices from a marginal or linear mixed-effects model.  The \strong{E} matrix is
calculated from the marginal variance-covariance (\strong{V}) matrix of an
unstructured covariance model fit with maximum likelihood.  The \strong{E} matrix is
used to calculate multivariate tests for MANOVA, univariate tests for
RM ANOVA, sphericity tests for within-subject effects, and epsilon
corrections for violations of sphericity.
}
\details{
The sum of squares cross products (SSCP) error matrices are
calculated as

\strong{E = M'} \eqn{(}\strong{Y'Y}\eqn{-} \strong{\eqn{\hat{B}}'} \eqn{(}\strong{X'X}\eqn{)}\strong{\eqn{\hat{B}}} \eqn{)}\strong{M}

where \strong{M} is an identity matrix for between-subject effects and a
sum-to-zero contrast matrix for within-subject effects, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, and \strong{X} is the design matrix of the
independent variables.

For marginal and mixed-effects models, the SSCP \strong{E} matrix is defined as

\strong{E =} \eqn{n}\strong{M'VM}

where \eqn{n} is the number of subjects, \strong{M} is an identity matrix for
between-subject effects and a sum-to-zero contrast matrix for within-subject
effects, and \strong{V} is the
marginal variance-covariance matrix where

\strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is the
corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix) for a
\code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}
unstructured covariance model object (refit with maximum likelihood instead
of restricted maximum likelihood).
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)

# SSCP between-subject E matrix
Ematrix(glsMod)$EB

# SSCP between-subject E matrix
Ematrix(manovaMod)$EB

# SSCP within-subject E matrix
Ematrix(glsMod)$E

# SSCP within-subject E matrix
Ematrix(manovaMod)$E

}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}

{\url{https://www.lesahoffman.com/PSYC943/mv12psyc943_lecture13.pdf}}
}
\seealso{
{\code{\link[mmemmuris]{coefs}}, \code{\link[mmemmuris]{Vmatrix}}}
}
