% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{epsilon}
\alias{epsilon}
\title{Epsilon Corrections for a Departure from Sphericity}
\usage{
epsilon(
  Ematrices,
  method = c("Greenhouse-Geisser", "Huynh-Feldt-Lecoutre"),
  tolerance = .Machine$double.eps
)
}
\arguments{
\item{Ematrices}{An object of class \code{Ematrix.ulm} or \code{Ematrix.mlm}.}

\item{method}{A character vector of which epsilon correction should be used.}
}
\value{
This function returns an epsilon value based on which correction was used.
}
\description{
This function will calculate Greenhouse-Geisser or Huynh-Feldt-Lecoutre
epsilons for univariate tests to correct within-subject effects for a
departure from sphericity for objects of class \code{\link[nlme]{gls}},
\code{\link[nlme]{lme}}, or \code{mlm}.
}
\details{
The sum of squares cross products (SSCP) error matrices are
calculated as

\strong{E = M'} \eqn{(}\strong{Y'Y}\eqn{-} \strong{\eqn{\hat{B}}'} \eqn{(}\strong{X'X}\eqn{)}\strong{\eqn{\hat{B}}} \eqn{)}\strong{M}

where \strong{M} is an identity matrix for between-subject effects and a
sum-to-zero contrast matrix for within-subject effects, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, and \strong{X} is the design matrix of the
independent variables.

For marginal and mixed-effects models, the SSCP \strong{E} matrix is defined as

\strong{E =} \eqn{n}\strong{M'VM}

where \eqn{n} is the number of subjects, \strong{M} is an identity matrix for
between-subject effects and a sum-to-zero contrast matrix for within-subject
effects, and \strong{V} is the
marginal variance-covariance matrix where

\strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is the
corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix) for a
\code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}
unstructured covariance model object (refit with maximum likelihood instead
of restricted maximum likelihood).

The within-subject SSCP \strong{E} matrix is used to calculate the
Greenhouse-Geisser and Huynh-Feldt-Lecoutre epsilons for univariate
adjustments for violations of sphericity.

The Greenhouse-Geisser epsilon is calculated as

\eqn{\epsilon_{GG} = ((\Sigma\lambda_i \text{ of }} \strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1}) ^ 2) / (r - 1)(\Sigma\lambda_i ^ 2 \text{ of }} \strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1})}

The Huynh-Feldt-Lecoutre epsilon is calculated as

\eqn{\epsilon_{HFL} = ((v + 1)(r - 1)\epsilon_{GG} - 2) / ((r - 1)(v - (r - 1)\epsilon_{GG}))}
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)

# Greenhouse-Geisser epsilon
epsilon(Ematrix(glsMod))
# Huynh-Feldt-Lecoutre epsilon
epsilon(Ematrix(glsMod), method = "H")

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)

# Greenhouse-Geisser epsilon
epsilon(Ematrix(manovaMod))
# Huynh-Feldt-Lecoutre epsilon
epsilon(Ematrix(manovaMod), method = "H")

}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}

{\url{https://www.lesahoffman.com/PSYC943/mv12psyc943_lecture13.pdf}}

{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_glm_details46.htm}}
}
\seealso{
{\code{\link[mmemmuris]{coefs}}, \code{\link[mmemmuris]{Vmatrix}}, \code{\link[mmemmuris]{Ematrix}}, \code{\link[mmemmuris]{sphericityTests}}}
}
