% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{hlTrace}
\alias{hlTrace}
\title{Multivariate Tests: Hotelling-Lawley Trace}
\usage{
hlTrace(
  fMod,
  individual = NULL,
  approximation = c("McKeon", "Pillai-Samson", "Wald"),
  within = "Time",
  waldF = TRUE,
  ss = NULL
)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm}.}

\item{individual}{The subject whose marginal variance-covariance matrix
should be used, where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design
matrix, \strong{G} is the corresponding (co)variances of the random effects matrix,
and \strong{R} represents the (co)variances of the repeated effects matrix).  The
default is the first subject in the factor.}

\item{approximation}{Approximation used for Hotelling-Lawley trace.  Options are
McKeon F ("McKeon"), Pillai-Samson F ("Pillai-Samson"), and Wald \eqn{\chi ^ 2}
("Wald").}

\item{within}{The name that should be used for the within-subject effect.
The default is "Time".}

\item{waldF}{Wald F-tests are used to calculate the Wald \eqn{\chi ^ 2} and
Hotelling-Lawley trace statistics instead of the \strong{H} and \strong{E} matrices.}
}
\value{
This function will return two lists: a within-subject effects list
(\code{withinTests}) and a between-subject effects list (\code{betweenTests}) if
between-subject effects are in the model.

\code{withinTests} is a list containing: \tabular{ll}{
\code{hlt}  \tab A \code{data.frame} containing the Hotelling-Lawley Trace \eqn{U} statistics and
hypothesis tests for these statistics. \cr
\code{parmsHL}  \tab A \code{data.frame} containing the values used to calculate
the Hotelling-Lawley Trace \eqn{U} statistics. \cr
\code{approximation}  \tab Approximation used for the Hotelling-Lawley Trace \eqn{U} statistics.  Options are McKeon F ("McKeon"),
Pillai-Samson F ("Pillai-Samson"), and Wald \eqn{\chi ^ 2} ("Wald"). \cr
\code{E}  \tab The sums of squares and crossproducts error matrix for within-subject effects. \cr
\code{H}  \tab The sums of squares and crossproducts hypothesis matrices for within-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\code{type}  \tab The type of \strong{L} matrix used.  Type 1 is sequential. \cr
\tab \cr
}
\code{betweenTests} is a list containing: \tabular{ll}{
\code{hlt}  \tab A \code{data.frame} containing the Hotelling-Lawley Trace \eqn{U} statistics and
hypothesis tests for these statistics. \cr
\code{parmsHL}  \tab A \code{data.frame} containing the values used to calculate
the Hotelling-Lawley Trace \eqn{U} statistics. \cr
\code{approximation}  \tab Approximation used for the Hotelling-Lawley Trace \eqn{U} statistics.  Options are McKeon F ("McKeon"),
Pillai-Samson F ("Pillai-Samson"), and Wald \eqn{\chi ^ 2} ("Wald"). \cr
\code{EB}  \tab The sums of squares and crossproducts error matrix for between-subject effects. \cr
\code{HB}  \tab The sums of squares and crossproducts hypothesis matrices for between-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\code{type}  \tab The type of \strong{L} matrix used.  Type 1 is sequential. \cr
\code{Vmatrix}  \tab The marginal variance-covariance matrix of the random/repeated effects. \cr
\tab \cr
}
}
\description{
This function will calculate Hotelling-Lawley Trace \eqn{U} statistics from marginal models,
mixed-effects models, and multivariate linear models.
}
\details{
\strong{H and E matrices}

The Hotelling-Lawley trace statistic for RM ANOVA is calculated as

\eqn{U = trace(}\strong{E}\eqn{ ^ {-1}}\strong{H}\eqn{)}

where the SSCP error matrix is

\strong{E = M'} \eqn{(}\strong{Y'Y}\eqn{-} \strong{\eqn{\hat{B}}'} \eqn{(}\strong{X'X}\eqn{)}\strong{\eqn{\hat{B}}} \eqn{)}\strong{M}

and the SSCP hypothesis matrix is

\strong{H} \eqn{=} \strong{M'} \eqn{(}\strong{L\eqn{\hat{B}}} \eqn{)}'\eqn{(}\strong{L}\eqn{(}\strong{X'X}\eqn{) ^ {-}}\strong{L'} \eqn{) ^ {-1}}\strong{L\eqn{\hat{B}}M}

where \strong{M} is an identity matrix for between-subject effects and a
sum-to-zero contrast matrix for within-subject effects, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, and \strong{X} is the design matrix of the
independent variables.

For marginal and mixed effects models, the \strong{E} matrix is calculated from
the marginal variance-covariance matrix of the random/repeated effects
(\strong{V})

\strong{E =} \eqn{n}\strong{M'VM}

where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is
the corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix).  The \strong{H}
matrix is calculated with the same formula as above.  The \strong{\eqn{\hat{B}}} matrix is
calculated by switching the reference levels of the within-subject effect,
refitting the models, and extracting the model coefficients.

\strong{Wald}

For marginal and mixed-effects models, we start with a Wald-type quadratic
form

\eqn{F = (}\strong{\eqn{\hat{\beta}}'L}\eqn{(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{) ^ {-1}}\strong{L'\eqn{\hat{\beta}}} \eqn{) / rank(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{)}

where \strong{\eqn{\hat{\beta}}} is a vector of fixed effects coefficients, \strong{X} is the fixed effects design matrix, and
\strong{V} is the marginal variance-covariance matrix
where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is
the corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix).  The contrast coefficients matrix (\strong{L}) is calculated from the \strong{LU}
decomposition of the crossproducts matrix (\strong{X'X}).

It follows that the Wald-\eqn{\chi ^ 2} statistic is the numerator degrees of
freedom (\eqn{rank(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{)})
multiplied by the F-statistic.  The Hotelling-Lawley trace (\eqn{U})
is then calculated as

\eqn{U = \chi ^ 2 / (n - rank(}\strong{X_b} \eqn{))}.

Let \eqn{v} be the residual degrees of freedom, \eqn{p = rank(}\strong{H}\eqn{+}\strong{E}\eqn{)},
\eqn{q = rank(}\strong{L}\eqn{(}\strong{X'X}\eqn{)^{-1}}\strong{L'} \eqn{)}, \eqn{s = min(p, q)}, \eqn{m = (|p - q| - 1) / 2} and \eqn{n = (v - p - 1) / 2}.

There are two approximations used for marginal and mixed-effects models:
Pillai-Samson (1959) and McKeon (1974).  The Pillai-Samson approximation is recommended for
\eqn{n \le 0} while the McKeon approximation is only defined for \eqn{n > 0}.
The McKeon approximation is said to be more accurate overall.

\strong{McKeon (1974)}

The McKeon statistic approximately follows an F-distribution where
\deqn{F_U = (U / c)((4 + (pq + 2) / (b - 1)) / (pq))}
with \eqn{pq} numerator degrees of freedom and \eqn{4 + (pq + 2) / (b - 1)}
denominator degrees of freedom, where

\eqn{b = (p + 2n)(q + 2n) / (2(2n + 1)(n - 1))} and

\eqn{c = (2 + (pq + 2) / (b - 1)) / (2n)}

If \eqn{s = 1}, the F-statistic is exact.

\strong{Pillai-Samson (1959)}

The Pillai-Samson statistic approximately follows an F-distribution where
\deqn{F_U = (2(sn + 1)U) / (s ^ 2(2m + s + 1))} with \eqn{s(2m + s + 1)}
numerator degrees of freedom and \eqn{2(sn + 1)} denominator degrees of
freedom.

If \eqn{s = 1}, the F-statistic is exact.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
# Chi-square statistic
hlTrace(glsMod, approximation = "Wald")
# Pillai-Samson F statistic
hlTrace(glsMod, approximation = "P")
# McKeon F statistic
hlTrace(glsMod)

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
# Chi-square statistic
hlTrace(lmeMod, approximation = "Wald")
# Pillai-Samson F statistic
hlTrace(lmeMod, approximation = "P")
# McKeon F statistic
hlTrace(lmeMod)

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)
# McKeon F statistic
hlTrace(manovaMod)
# Pillai-Samson F statistic
hlTrace(manovaMod, approximation = "P")
# Chi-square statistic
hlTrace(manovaMod, approximation = "Wald")

}
\references{
{\url{https://support.sas.com/resources/papers/proceedings/proceedings/sugi23/Stats/p229.pdf}}

{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}

{McKeon, J. J. (1974). "F Approximations to the Distribution of Hotelling's \eqn{T_0^2}." Biometrika 61:381-383.}

{Pillai, K. C. S., and Samson, P., Jr. (1959). "On Hotelling's Generalization of \eqn{T_0^2}." Biometrika 46:160-168.}

{Wright, S. P. (1994). Adjusted F Tests for Repeated Measures with
the MIXED Procedure. Knoxville: Statistics Department, University of Tennessee.
\url{https://support.sas.com/resources/papers/proceedings-archive/SUGI95/Sugi-95-195\%20Wright.pdf}}

{\url{https://brainder.org/tag/lawley-hotelling-trace/}}
}
\seealso{
{\code{\link[mmemmuris]{Hmatrix}}, \code{\link[mmemmuris]{coefs}}, \code{\link[mmemmuris]{Vmatrix}}, \code{\link[mmemmuris]{Ematrix}}, \code{\link[mmemmuris]{waldF}}, \code{\link[mmemmuris]{MLtoREML}}}
}
