% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{Hmatrix}
\alias{Hmatrix}
\title{MANOVA/RM ANOVA SSCP Hypothesis (\strong{H}) Matrix}
\usage{
Hmatrix(fMod, within = "Time")
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm}.}

\item{within}{The name that should be used for the within-subject effect.
The default is "Time".}
}
\value{
\tabular{ll}{
\code{H}  \tab The sums of squares and crossproducts hypothesis matrices for within-subject effects. \cr
\code{M} \tab A sum-to-zero contrast matrix for within-subject effects. \cr
\code{SSH} \tab The sums of squares of the hypothesis matrices for within-subject effects. \cr
\code{HB}  \tab The sums of squares and crossproducts hypothesis matrices for between-subject effects. \cr
\code{MB} \tab An identity contrast matrix for between-subject effects. \cr
\code{B} \tab The multivariate linear model coefficients. \cr
\code{X} \tab The design matrix of the between-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\tab \cr
}
}
\description{
This function returns the hypothesis (\strong{H}) sums of squares and crossproducts (SSCP)
matrices from a marginal or linear mixed-effects model.  The \strong{H} matrix is
used to calculate multivariate tests for MANOVA and univariate tests for
RM ANOVA.
}
\details{
The sum of squares and crossproducts (SSCP) hypothesis matrix is

\strong{H} \eqn{=} \strong{M'} \eqn{(}\strong{L\eqn{\hat{B}}} \eqn{)}'\eqn{(}\strong{L}\eqn{(}\strong{X'X}\eqn{) ^ {-}}\strong{L'} \eqn{) ^ {-1}}\strong{L\eqn{\hat{B}}M}

where \strong{M} is a sum-to-zero contrast matrix, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, \strong{X} is the design matrix of the
independent variables, and the contrast coefficients matrix (\strong{L}) is
calculated from the \strong{LU} decomposition of the crossproducts matrix
(\strong{X'X}).

For marginal and mixed effects models, the \strong{H} matrix is calculated with
the same formula as above.  The \strong{\eqn{\hat{B}}} matrix is calculated by switching the
reference levels of the within-subject effect, refitting the models, and
extracting the model coefficients.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)

# SSCP between-subject H matrix
Hmatrix(glsMod)$HB

# SSCP between-subject H matrix
Hmatrix(manovaMod)$HB

# SSCP within-subject H matrix
Hmatrix(glsMod)$H

# SSCP within-subject H matrix
Hmatrix(manovaMod)$H
}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}
}
