% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{isReversible}
\alias{isReversible}
\title{Check If Marginal and Mixed-Effects Models Correspond to a Multivariate Model}
\usage{
isReversible(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}
}
\value{
This function returns a logical telling the user if the marginal or
mixed-effects model corresponds to a multivariate linear model.
}
\description{
This function checks if a marginal (\code{\link[nlme]{gls}}) or mixed-effects
(\code{\link[nlme]{lme}}) model is reversible.  A marginal or mixed-effects
model is reversible if it can be equivalently specified as a multivariate linear model.
}
\details{
A reversible model needs complete cases (no missing data or all cases with
missing data omitted), all between-subject terms interacted with the
within-subject terms, and the marginal variance-covariance matrix of the
random/repeated effects for each subject must be the same unstructured
covariance structure.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)

# Reversible model
isReversible(glsMod)

# Non-reversible model
isReversible(update(glsMod, model = Distance ~ Sex + Age))

}
\references{
{\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7009022/}}
}
\seealso{
{\code{\link[mmemmuris]{completeData}}, \code{\link[mmemmuris]{BWInteracted}}}
}
