% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{Lmatrix.ulm}
\alias{Lmatrix.ulm}
\title{Create the L coefficient contrast matrix}
\usage{
Lmatrix.ulm(fMod)
}
\description{
This function creates a Type 1 Sum of Squares \strong{L} matrix.
}
\details{
The contrast coefficients matrix (\strong{L}) is created from an \eqn{LU}
decomposition of the crossproduct matrix \strong{X'X} with the design matrix \strong{X}
such that
\deqn{PA = LU}

The \strong{L} matrix is then defined as

\strong{L} \eqn{= L'}
}
