% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{LRT}
\alias{LRT}
\title{Likelihood Ratio Test with a Contrast Coefficients Matrix}
\usage{
LRT(fMod, L)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}

\item{L}{A contrast coefficients matrix.}
}
\value{
This function will return a \code{data.frame} for each term in the model.  Each
\code{data.frame} contains \tabular{ll}{
\code{df} \tab The degrees of freedom used for the models. \cr
\tab \cr
\code{AIC} \tab The Akaike information criterions (AICs) for the models. \cr
\tab \cr
\code{BIC} \tab The Bayesian information criterions (BICs) for the models. \cr
\tab \cr
\code{logLik} \tab The log-likelihoods for the models. \cr
\tab \cr
\code{L.Ratio} \tab The likelihood ratio for the models. \cr
\tab \cr
\verb{Pr(>Chisq)} \tab The p-value for the likelihood ratio test for the models. \cr
}
}
\description{
This function performs a likelihood ratio test for a full and reduced model.
}
\details{
The maximum log-likelihood function for the full model is

\eqn{logLik(M_F) = (-1 / 2)log(det(}\strong{V_{F}} \eqn{)) - ((1 / 2)}\strong{r'V_{F}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

where \strong{V_{F}} is the marginal variance-covariance matrix of the random/repeated
effects, \strong{r} is the vector of model residuals, and \eqn{N} is the number of
observations of data in the long format.

The contrast coefficients matrix (\strong{L}) is calculated from the \strong{LU}
decomposition of the crossproducts matrix (\strong{X_{F}'X_{F}}).  To
project the design matrix of the reduced model (\strong{X_{R}}) onto the
column space of the design matrix of the full model, the design matrix of the
full model needs to be multiplied by the orthogonal complement of \strong{L'}.
It follows that the maximum log-likelihood function for the reduced model is

\eqn{logLik(M_R) = (-1 / 2)log(det(}\strong{V_{R}} \eqn{)) - ((1 / 2)}\strong{r'V_{R}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

where \strong{V_{R}} is the marginal variance-covariance matrix of the random/repeated
effects, \strong{r} is the vector of model residuals, and \eqn{N} is the number of
observations of data in the long format.

The likelihood ratio (LR) statistic approximately follows a \eqn{\chi ^ 2}
distribution \deqn{LR = -2logLik(M_R) - \{-2logLik(M_F)\}} with
\eqn{df_{M_R} - df_{M_F}} degrees of freedom.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)

# Likelihood ratio chi-square statistic
L <- mmemmuris:::Lmatrix.ulm(glsMod)
L$Sex <- rbind(L$Sex, L$`Sex:Age`)
LRT(glsMod, L)

# Likelihood ratio chi-square statistic
wilksLambda(glsMod, approximation = "LR")

}
\seealso{
{\code{\link[mmemmuris]{REMLtoML}}, \code{\link[mmemmuris]{wilksLambda}}}
}
