% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{MLtoREML}
\alias{MLtoREML}
\title{Refit Model with Restricted Maximum Likelihood}
\usage{
MLtoREML(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}
}
\value{
This function will return a model of class \code{\link[nlme]{gls}} or
\code{\link[nlme]{lme}} fit with restricted maximum likelihood.
}
\description{
Models fit with maximum likelihood are refit with restricted maximum
likelihood (REML).  In this package, models are fit with REML to ensure the
(co)variance parameter estimates are unbiased estimates.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong, method = "ML")

# Marginal variance-covariance matrix of the random/repeated effects for
# SSCP E matrix
Vmatrix(glsMod)$V

# Fit with REML for marginal variance-covariance matrix of the
# random/repeated effects for compound symmetry covariance
Vmatrix(MLtoREML(update(glsMod,
                        correlation = corCompSymm(form = ~ 1 | Subject),
                        weights = NULL)))
}
