% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{nestedCovariance}
\alias{nestedCovariance}
\title{Likelihood Ratio Test for Nonsymbolically Nested Models (Compound Symmetry Covariance Nested Within an Unstructured Covariance)}
\usage{
nestedCovariance(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}},
or \code{mlm}.}
}
\value{
\tabular{ll}{
\code{df} \tab The degrees of freedom used for the models. \cr
\tab \cr
\code{AIC} \tab The Akaike information criterions (AICs) for the models. \cr
\tab \cr
\code{BIC} \tab The Bayesian information criterions (BICs) for the models. \cr
\tab \cr
\code{logLik} \tab The log-likelihoods for the models. \cr
\tab \cr
\code{L.Ratio} \tab The likelihood ratio for the models. \cr
\tab \cr
\verb{Pr(>Chisq)} \tab The p-value for the likelihood ratio test for the models. \cr
}
}
\description{
This function will provide a possible rationale for choosing to use the
multivariate tests (unstructured) or univariate tests (compound symmetry)
based on a likelihood ratio test, AIC, or BIC of the covariance structures
for a \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm} model
object.
}
\details{
\strong{Marginal and Mixed-Effects Models}

The models with compound symmetry and unstructured covariance structures are
fit with maximum likelihood so that a likelihood ratio test can be done for
the covariance structures.  The maximum log-likelihood functions for the
models with compound symmetry and unstructured covariances, respectively, are

\eqn{logLik(M_R) = (-1 / 2)log(det(}\strong{V_{CS}} \eqn{)) - ((1 / 2)}\strong{r'V_{CS}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

and

\eqn{logLik(M_F) = (-1 / 2)log(det(}\strong{V_{UN}} \eqn{)) - ((1 / 2)}\strong{r'V_{UN}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

where \strong{V} is the marginal variance-covariance matrix of the random/repeated
effects, \strong{r} is the vector of model residuals, and \eqn{N} is the number of
observations of data in the long format.

The likelihood ratio (LR) statistic approximately follows a \eqn{\chi ^ 2}
distribution \deqn{LR = -2logLik(M_R) - {-2logLik(M_F)}} with
\eqn{df_{M_R} - df_{M_F}} degrees of freedom.

The Akaike's Information Criterion (AIC) is defined as
\deqn{AIC = 2df_M - 2logLik(M)}
where \eqn{df_M} is the number of parameter estimates in the model.

The Bayesian Information Criterion (BIC) is defined as
\deqn{BIC = df_Mlog(N) - 2logLik(M)}
where \eqn{df_M} is the number of parameter estimates in the model.

\strong{MANOVA/RM ANOVA}

For MANOVA/RM ANOVA, the (log-)likelihoods are not computed since method of
moments is used instead of (RE)ML.  Therefore, this could be viewed as a fake
solution used for illustration purposes.  We start by obtaining the marginal
variance-covariance matrix of the the random/repeated effects from the
between-subject effects \strong{E} matrix.  The model with an unstructured
covariance using method of moments coincides with the maximum likelihood
estimates for the marginal variance-covariance matrix.  However, the model
with a compound symmetry covariance using method of moments coincides with
the restricted maximum likelihood estimates for the marginal
variance-covariance matrix.  Since the maximum likelihood estimates are
biased, we need to undo Bessel's correction by multiplying the marginal
variance-covariance matrix "obtained" by REML by \eqn{v / n} where \eqn{v} is
the residual degrees of freedom for the multivariate linear model and \eqn{n}
is the total number of observations in the dataset in the wide format.

The maximum log-likelihood functions for the
models with compound symmetry and unstructured covariances, respectively, are

\eqn{logLik(M_R) = (-1 / 2)log(det(}\strong{V_{CS}} \eqn{)) - ((1 / 2)}\strong{r'V_{CS}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

and

\eqn{logLik(M_F) = (-1 / 2)log(det(}\strong{V_{UN}} \eqn{)) - ((1 / 2)}\strong{r'V_{UN}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

where \strong{V} is the marginal variance-covariance matrix of the random/repeated
effects, \strong{r} is the vector of model residuals, and \eqn{N} is the number of
observations of data in the long format.

The likelihood ratio (LR) statistic approximately follows a \eqn{\chi ^ 2}
distribution \deqn{LR = -2logLik(M_R) - {-2logLik(M_F)}} with
\eqn{df_{M_R} - df_{M_F}} degrees of freedom.

The Akaike's Information Criterion (AIC) is defined as
\deqn{AIC = 2df_M - 2logLik(M)}
where \eqn{df_M} is the number of parameter estimates in the model.

The Bayesian Information Criterion (BIC) is defined as
\deqn{BIC = df_Mlog(N) - 2logLik(M)}
where \eqn{df_M} is the number of parameter estimates in the model.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
nestedCovariance(glsMod)

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
nestedCovariance(lmeMod)

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)
nestedCovariance(manovaMod)

}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.3/statug/statug_mixed_details01.htm#statug.mixed.mixedmmtgandr}}
}
\seealso{
{\code{\link[mmemmuris]{covStruct}}, \code{\link[mmemmuris]{REMLtoML}}}
}
