% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{pillaiTrace}
\alias{pillaiTrace}
\title{Multivariate Tests: Pillai's Trace}
\usage{
pillaiTrace(
  fMod,
  individual = NULL,
  approximation = c("Muller", "Pillai"),
  within = "Time",
  LR = TRUE,
  ss = NULL
)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm}.}

\item{individual}{The subject whose marginal variance-covariance matrix
should be used, where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design
matrix, \strong{G} is the corresponding (co)variances of the random effects matrix,
and \strong{R} represents the (co)variances of the repeated effects matrix).  The
default is the first subject in the factor.}

\item{approximation}{Approximation used for Pillai's trace.  Options are
Pillai F ("Pillai") and Muller (Method #2) F ("Muller").}

\item{within}{The name that should be used for the within-subject effect.
The default is "Time".}

\item{LR}{Likelihood ratio tests are used to calculate the likelihood ratio
\eqn{\chi ^ 2} and Wilks' \eqn{\Lambda} statistics instead of the \strong{H} and
\strong{E} matrices.}
}
\value{
This function will return two lists: a within-subject effects list
(\code{withinTests}) and a between-subject effects list (\code{betweenTests}) if
between-subject effects are in the model.

\code{withinTests} is a list containing: \tabular{ll}{
\code{pillai}  \tab A \code{data.frame} containing the Pillai's Trace \eqn{V} statistics and
hypothesis tests for these statistics. \cr
\code{parmsPillai}  \tab A \code{data.frame} containing the values used to calculate
the Pillai's Trace \eqn{V} statistics. \cr
\code{approximation}  \tab Approximation used for the Pillai's Trace \eqn{V} statistics.  Options are Muller F ("Muller") and
and Pillai F ("Pillai"). \cr
\code{E}  \tab The sums of squares and crossproducts error matrix for within-subject effects. \cr
\code{H}  \tab The sums of squares and crossproducts hypothesis matrices for within-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\code{type}  \tab The type of \strong{L} matrix used.  Type 1 is sequential. \cr
\tab \cr
}
\code{betweenTests} is a list containing: \tabular{ll}{
\code{pillai}  \tab A \code{data.frame} containing the Pillai's Trace \eqn{V} statistics and
hypothesis tests for these statistics. \cr
\code{parmsPillai}  \tab A \code{data.frame} containing the values used to calculate
the Pillai's Trace \eqn{V} statistics. \cr
\code{approximation}  \tab Approximation used for the Pillai's Trace \eqn{V} statistics.  Options are Muller F ("Muller") and
and Pillai F ("Pillai"). \cr
\code{EB}  \tab The sums of squares and crossproducts error matrix for between-subject effects. \cr
\code{HB}  \tab The sums of squares and crossproducts hypothesis matrices for between-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\code{type}  \tab The type of \strong{L} matrix used.  Type 1 is sequential. \cr
\code{Vmatrix}  \tab The marginal variance-covariance matrix of the random/repeated effects. \cr
\tab \cr
}
}
\description{
This function will calculate Pillai's Trace \eqn{V} statistics from marginal models,
mixed-effects models, and multivariate linear models.
}
\details{
\strong{H and E matrices}

The Pillai's trace statistic for a RM ANOVA is denoted as

\eqn{V = trace(}\strong{H}\eqn{(}\strong{H}\eqn{+}\strong{E}\eqn{)^{-1})}

where the SSCP error matrix is

\strong{E = M'} \eqn{(}\strong{Y'Y}\eqn{-} \strong{\eqn{\hat{B}}'} \eqn{(}\strong{X'X}\eqn{)}\strong{\eqn{\hat{B}}} \eqn{)}\strong{M}

and the SSCP hypothesis matrix is

\strong{H} \eqn{=} \strong{M'} \eqn{(}\strong{L\eqn{\hat{B}}} \eqn{)}'\eqn{(}\strong{L}\eqn{(}\strong{X'X}\eqn{) ^ {-}}\strong{L'} \eqn{) ^ {-1}}\strong{L\eqn{\hat{B}}M}

where \strong{M} is an identity matrix for between-subject effects and a
sum-to-zero contrast matrix for within-subject effects, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, and \strong{X} is the design matrix of the
independent variables.

For marginal and mixed effects models, the \strong{E} matrix is calculated from
the marginal variance-covariance matrix of the random/repeated effects
(\strong{V})

\strong{E =} \eqn{n}\strong{M'VM}

where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is
the corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix).  The \strong{H}
matrix is calculated with the same formula as above.  The \strong{\eqn{\hat{B}}} matrix is
calculated by switching the reference levels of the within-subject effect,
refitting the models, and extracting the model coefficients.

\strong{Likelihood Ratio}

The Wilks' lambda statistic for a marginal or mixed-effects model is
essentially a likelihood ratio test with the reduced model excluding the
term being tested.  The log-likelihood function for the full model is

\eqn{logLik(M_F) = (-1 / 2)log(det(}\strong{V_{F}} \eqn{)) - ((1 / 2)}\strong{r'V_{F}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

where \strong{V_{F}} is the marginal variance-covariance matrix of the random/repeated
effects, \strong{r} is the vector of model residuals, and \eqn{N} is the number of
observations of data in the long format.

The contrast coefficients matrix (\strong{L}) is calculated from the \strong{LU}
decomposition of the crossproducts matrix (\strong{X_{F}'X_{F}}).  To
project the design matrix of the reduced model (\strong{X_{R}}) onto the
column space of the design matrix of the full model, the design matrix of the
full model needs to be multiplied by the orthogonal complement of \strong{L'}.
It follows that the log-likelihood function for the reduced model is

\eqn{logLik(M_R) = (-1 / 2)log(det(}\strong{V_{R}} \eqn{)) - ((1 / 2)}\strong{r'V_{R}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

where \strong{V_{R}} is the marginal variance-covariance matrix of the random/repeated
effects, \strong{r} is the vector of model residuals, and \eqn{N} is the number of
observations of data in the long format.

Wilks' \eqn{\Lambda} is denoted as
\deqn{\Lambda = exp{(-[-2logLik(M_R) - {-2logLik(M_F)}]) / n}}
where \eqn{n} is the number of subjects.

For marginal and mixed-effects models, if \eqn{s = min(p, q) = 1}, Pillai's
trace is calculated as \eqn{V = 1 - \Lambda}.  With \code{LR = TRUE},
Pillai's trace is only calculated when \eqn{s = 1} for within-subject
effects.

\strong{Pillai (1954)}

Let \eqn{v} be the residual degrees of freedom, \eqn{p = rank(}\strong{H}\eqn{+}\strong{E}\eqn{)},
\eqn{q = rank(}\strong{L}\eqn{(}\strong{X'X}\eqn{)^{-1}}\strong{L'} \eqn{)}, \eqn{s = min(p, q)},
\eqn{m = (|p - q| - 1) / 2} and \eqn{n = (v - p - 1) / 2}.  The
statistic approximately follows an F-distribution
\deqn{F_V = ((2n + s + 1) / (2m + s + 1))(V / (s - V))} with
\eqn{s(2m + s + 1)} numerator degrees of freedom and \eqn{s(2n + s + 1)}
denominator degrees of freedom.

If \eqn{s = 1}, the F-statistic is exact.

\strong{Muller Method #2 (1998)}

Let \eqn{v} be the residual degrees of freedom, \eqn{p = rank(}\strong{H}\eqn{+}\strong{E}\eqn{)},
\eqn{q = rank(}\strong{L}\eqn{(}\strong{X'X}\eqn{)^{-1}}\strong{L'} \eqn{)}, and \eqn{s = min(p, q)}.  The
statistic approximately follows an F-distribution
\deqn{F_V = ((((v + s - p) / (v + q))[(((s(v + s - p)(v + q + 2)(v + q - 1)) / (v(v + q - p))) - 2)]) / (((qp) / (s(v + q)))[(((s(v + s - p)(v + q + 2)(v + q - 1)) / (v(v + q - p))) - 2)]))(V / (s - V))} with
\eqn{((qp) / (s(v + q)))[(((s(v + s - p)(v + q + 2)(v + q - 1)) / (v(v + q - p))) - 2)]} numerator degrees of freedom and \eqn{(((v + s - p) / (v + q))[(((s(v + s - p)(v + q + 2)(v + q - 1)) / (v(v + q - p))) - 2)])}
denominator degrees of freedom.

If \eqn{s = 1}, the F-statistic is exact.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
# Pillai F statistic
pillaiTrace(glsMod)

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
# Pillai F statistic
pillaiTrace(lmeMod)

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)
# Pillai F statistic
pillaiTrace(manovaMod)
# Muller F statistic (Method #2)
pillaiTrace(manovaMod, approximation = "M")

}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}

{Muller, K. E. (1998). "A New F Approximation for the Pillai-Bartlett Trace under \eqn{H_0}." Journal of Computational and Graphical Statistics 7:131-137.}
}
\seealso{
{\code{\link[mmemmuris]{Hmatrix}}, \code{\link[mmemmuris]{coefs}}, \code{\link[mmemmuris]{Vmatrix}}, \code{\link[mmemmuris]{Ematrix}}, \code{\link[mmemmuris]{REMLtoML}}, \code{\link[mmemmuris]{LRT}}}
}
