% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{print.nestedCovariance}
\alias{print.nestedCovariance}
\title{Print a Nested Covariance Object}
\usage{
\method{print}{nestedCovariance}(nestedCovariance, digits = 4)
}
\description{
Print a Nested Covariance Object
}
