% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{print.sphericityTests}
\alias{print.sphericityTests}
\title{Print a Sphericity Tests Object}
\usage{
\method{print}{sphericityTests}(sphericityTests, digits = 4)
}
\description{
Print a Sphericity Tests Object
}
