% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{print.uniTest}
\alias{print.uniTest}
\title{Print a Univariate Tests Object}
\usage{
\method{print}{uniTest}(uniTest, V = FALSE, H = FALSE, E = FALSE, digits = 4)
}
\description{
Print a Univariate Tests Object
}
