% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{REMLtoML}
\alias{REMLtoML}
\title{Refit Model with Maximum Likelihood}
\usage{
REMLtoML(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}
}
\value{
This function will return a model of class \code{\link[nlme]{gls}} or
\code{\link[nlme]{lme}} fit with maximum likelihood.
}
\description{
Models fit with restricted maximum likelihood are refit with maximum
likelihood (ML).  In this package, models are fit with ML for likelihood
ratio tests and to calculate the error sum of squares and crossproducts
matrix (\strong{E}).
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)

# Likelihood ratio test
anova(REMLtoML(update(glsMod, model = Distance ~ Sex + Age)),
      REMLtoML(glsMod))
wilksLambda(glsMod, approximation = "LR")$withinTests$wilks[2, ]

# SSCP E matrix
27 * Vmatrix(REMLtoML(glsMod))$V
Ematrix(glsMod)$EB
}
