% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{royGR}
\alias{royGR}
\title{Multivariate Tests: Roy's Greatest Root}
\usage{
royGR(fMod, individual = NULL, within = "Time", waldF = TRUE, ss = NULL)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}

\item{individual}{The subject whose marginal variance-covariance matrix
should be used, where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design
matrix, \strong{G} is the corresponding (co)variances of the random effects matrix,
and \strong{R} represents the (co)variances of the repeated effects matrix).  The
default is the first subject in the factor.}

\item{within}{The name that should be used for the within-subject effect.
The default is "Time".}

\item{waldF}{Wald F-tests are used to calculate the Wald \eqn{\chi ^ 2} and
Hotelling-Lawley trace statistics instead of the \strong{H} and \strong{E} matrices.}
}
\value{
This function will return two lists: a within-subject effects list
(\code{withinTests}) and a between-subject effects list (\code{betweenTests}) if
between-subject effects are in the model.

\code{withinTests} is a list containing: \tabular{ll}{
\code{rgr}  \tab A \code{data.frame} containing the Roy's Greatest Root \eqn{\Theta} statistics and
hypothesis tests for these statistics. \cr
\code{parmsRGR}  \tab A \code{data.frame} containing the values used to calculate
the Roy's Greatest Root \eqn{\Theta} statistics. \cr
\code{E}  \tab The sums of squares and crossproducts error matrix for within-subject effects. \cr
\code{H}  \tab The sums of squares and crossproducts hypothesis matrices for within-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\code{type}  \tab The type of \strong{L} matrix used.  Type 1 is sequential. \cr
\tab \cr
}
\code{betweenTests} is a list containing: \tabular{ll}{
\code{rgr}  \tab A \code{data.frame} containing the Roy's Greatest Root \eqn{\Theta} statistics and
hypothesis tests for these statistics. \cr
\code{parmsRGR}  \tab A \code{data.frame} containing the values used to calculate
the Roy's Greatest Root \eqn{\Theta} statistics. \cr
\code{EB}  \tab The sums of squares and crossproducts error matrix for between-subject effects. \cr
\code{HB}  \tab The sums of squares and crossproducts hypothesis matrices for between-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\code{type}  \tab The type of \strong{L} matrix used.  Type 1 is sequential. \cr
\code{Vmatrix}  \tab The marginal variance-covariance matrix of the random/repeated effects. \cr
\tab \cr
}
}
\description{
This function will calculate Roy's Greatest Root \eqn{\Theta} statistics from marginal models,
mixed-effects models, and multivariate linear models.
}
\details{
The Roy's greatest root statistic for RM ANOVA is calculated as

\eqn{\Theta = max(\lambda_i)} of \strong{E}\eqn{ ^ {-1}}\strong{H}

where the SSCP error matrix is

\strong{E = M'} \eqn{(}\strong{Y'Y}\eqn{-} \strong{\eqn{\hat{B}}'} \eqn{(}\strong{X'X}\eqn{)}\strong{\eqn{\hat{B}}} \eqn{)}\strong{M}

and the SSCP hypothesis matrix is

\strong{H} \eqn{=} \strong{M'} \eqn{(}\strong{L\eqn{\hat{B}}} \eqn{)}'\eqn{(}\strong{L}\eqn{(}\strong{X'X}\eqn{) ^ {-}}\strong{L'} \eqn{) ^ {-1}}\strong{L\eqn{\hat{B}}M}

where \strong{M} is an identity matrix for between-subject effects and a
sum-to-zero contrast matrix for within-subject effects, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, and \strong{X} is the design matrix of the
independent variables.

For marginal and mixed effects models, the \strong{E} matrix is calculated from
the marginal variance-covariance matrix of the random/repeated effects
(\strong{V})

\strong{E =} \eqn{n}\strong{M'VM}

where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is
the corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix).  The \strong{H}
matrix is calculated with the same formula as above.  The \strong{\eqn{\hat{B}}} matrix is
calculated by switching the reference levels of the within-subject effect,
refitting the models, and extracting the model coefficients.

\strong{Wald}

For marginal and mixed-effects models, we start with a Wald-type quadratic
form

\eqn{F = (}\strong{\eqn{\hat{\beta}}'L}\eqn{(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{) ^ {-1}}\strong{L'\eqn{\hat{\beta}}} \eqn{) / rank(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{)}

where \strong{\eqn{\hat{\beta}}} is a vector of fixed effects coefficients, \strong{X} is the fixed effects design matrix, and
\strong{V} is the marginal variance-covariance matrix
where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is
the corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix).  The contrast coefficients matrix (\strong{L}) is calculated from the \strong{LU}
decomposition of the crossproducts matrix (\strong{X'X}).

It follows that the Wald-\eqn{\chi ^ 2} statistic is the numerator degrees of
freedom (\eqn{rank(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{)})
multiplied by the F-statistic.  The Hotelling-Lawley trace (\eqn{U})
is then calculated as

\eqn{U = \chi ^ 2 / (n - rank(}\strong{X_b} \eqn{))}.

For marginal and mixed-effects models, if \eqn{s = min(p, q) = 1}, Roy's
greatest root is calculated as \eqn{\Theta = U}.  With \code{waldF = TRUE},
Roy's greatest root is only calculated when \eqn{s = 1} for within-subject
effects.

Let \eqn{v} be the residual degrees of freedom, \eqn{p = rank(}\strong{H}\eqn{+}\strong{E}\eqn{)},
\eqn{q = rank(}\strong{L}\eqn{(}\strong{X'X}\eqn{)^{-1}}\strong{L'} \eqn{)}, and \eqn{s = min(p, q)}.

The statistic approximately follows an F-distribution
\deqn{F_\Theta = \Theta((v - max(p, q) + q) / max(p, q))}
with \eqn{max(p, q)} numerator degrees of freedom and \eqn{v - max(p, q) + q}
denominator degrees of freedom.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
# F statistic
royGR(glsMod)

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
# F statistic
royGR(lmeMod)

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)
# F statistic
royGR(manovaMod)

}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}
}
\seealso{
{\code{\link[mmemmuris]{Hmatrix}}, \code{\link[mmemmuris]{coefs}}, \code{\link[mmemmuris]{Vmatrix}}, \code{\link[mmemmuris]{Ematrix}}, \code{\link[mmemmuris]{waldF}}, \code{\link[mmemmuris]{MLtoREML}}}
}
