% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{sigStars}
\alias{sigStars}
\title{Significance Level Stars for p-values}
\usage{
sigStars(pValues)
}
\arguments{
\item{pValues}{A vector of p-values.}
}
\value{
This function returns a vector of significance stars.
}
\description{
This creates a vector of significance level stars for p-values.

Significance level codes: \tabular{lll}{ \cr
[0, 0.001]    \tab - \tab *** \cr
(0.001, 0.01] \tab - \tab ** \cr
(0.01, 0.05]  \tab - \tab * \cr
(0.05, 0.1]   \tab - \tab . \cr
(0.1, 1]      \tab -
}
}
\examples{
pValues <- c(0, 1e-16, 0.005, 0.025, 0.075, 0.5, 1)
sigStars(pValues)
}
