% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{sphericityTests}
\alias{sphericityTests}
\title{Mauchly's Sphericity Tests for Within-Subject Effects}
\usage{
sphericityTests(Ematrices)
}
\arguments{
\item{Ematrices}{An object of class \code{Ematrix.ulm} or \code{Ematrix.mlm}.}
}
\value{
\tabular{ll}{
\code{sphericityTests}  \tab A \code{data.frame} containing the Mauchly's sphericity
test statistics. \cr
\code{r}  \tab The total number of repeated measurements for data in the long format. \cr
\code{v}  \tab The residual degrees of freedom. \cr
\code{E}  \tab The sums of squares and crossproducts error matrix for within-subject effects. \cr
\code{M}  \tab A sum-to-zero contrast matrix for within-subject effects. \cr
\tab \cr
}
}
\description{
This function will calculate Mauchly's test of sphericity for within-subject
effects for marginal models (\code{\link[nlme]{gls}}), mixed-effects models
(\code{\link[nlme]{lme}}), and multivariate linear models (\code{mlm}).
}
\details{
The sum of squares cross products (SSCP) error matrices are
calculated as

\strong{E = M'} \eqn{(}\strong{Y'Y}\eqn{-} \strong{\eqn{\hat{B}}'} \eqn{(}\strong{X'X}\eqn{)}\strong{\eqn{\hat{B}}} \eqn{)}\strong{M}

where \strong{M} is an identity matrix for between-subject effects and a
sum-to-zero contrast matrix for within-subject effects, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, and \strong{X} is the design matrix of the
independent variables.

For marginal and mixed-effects models, the SSCP \strong{E} matrix is defined as

\strong{E =} \eqn{n}\strong{M'VM}

where \eqn{n} is the number of subjects, \strong{M} is an identity matrix for
between-subject effects and a sum-to-zero contrast matrix for within-subject
effects, and \strong{V} is the
marginal variance-covariance matrix where

\strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is the
corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix) for a
\code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}
unstructured covariance model object (refit with maximum likelihood instead
of restricted maximum likelihood).

The within-subject SSCP \strong{E} matrix is used to calculate Mauchly's test for
sphericity for within-subject effects.

Mauchly's test for sphericity is calculated as

\eqn{W = (\Pi\lambda_i of} \strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1}) / (((\Sigma\lambda_i of} \strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1}) / (r - 1)) ^ (r - 1))}

Let \eqn{r} be equal to the number of repeated measurements, \eqn{v}
be the residual degrees of freedom (between-subject design matrix), and the correction factor equal
\deqn{\rho = 1 - (2(r - 1) ^ 2 + r + 1) / (6(r - 1)v)}  The \eqn{\chi ^ 2} statistic is
equal to the product \eqn{-\rho v log(W)} with degrees of freedom
\eqn{(r(r - 1) / 2) - 1}.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
sphericityTests(Ematrix(glsMod))

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
sphericityTests(Ematrix(lmeMod))

# Multivariate linear model
manovaMod <-
  manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
         data = orthodontWide)
sphericityTests(Ematrix(manovaMod))

}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}

{\url{https://www.lesahoffman.com/PSYC943/mv12psyc943_lecture13.pdf}}

{\url{https://support.sas.com/rnd/app/stat/papers/mixedglm.pdf}}
}
\seealso{
{\code{\link[mmemmuris]{coefs}}, \code{\link[mmemmuris]{Vmatrix}}, \code{\link[mmemmuris]{Ematrix}}, \code{\link[mmemmuris]{epsilon}}}
}
