% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{termsType}
\alias{termsType}
\title{Terms Type}
\usage{
termsType(fMod)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}
}
\value{
\tabular{ll}{
\code{between} \tab A character vector of between-subject effects. \cr
\tab \cr
\code{within} \tab A character vector of within-subject effects. \cr
\tab \cr
\code{Xbetween} \tab A sum-to-zero design matrix of between-subject effects.
\cr \tab \cr
}
}
\description{
This function will determine the between and within-subject effects from a
\code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}
model object.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
termsType(glsMod)

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
termsType(lmeMod)
}
