% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{uniTest}
\alias{uniTest}
\title{Univariate Tests for RM ANOVA, Mixed-Effects Models, or Marginal Models}
\usage{
uniTest(
  fMod,
  individual = NULL,
  within = "Time",
  epsilon = c("Greenhouse-Geisser", "Huynh-Feldt-Lecoutre"),
  refit = TRUE,
  ss = NULL
)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm}.}

\item{individual}{The subject whose marginal variance-covariance matrix
should be used, where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design
matrix, \strong{G} is the corresponding (co)variances of the random effects matrix,
and \strong{R} represents the (co)variances of the repeated effects matrix).  The
default is the first subject in the factor.}

\item{within}{The name that should be used for the within-subject effect.
The default is "Time".}

\item{epsilon}{Epsilon degrees of freedom adjustment used.  Options are
"Greenhouse-Geisser" and "Huynh-Feldt-Lecoutre".}

\item{refit}{The marginal or mixed-effects model is refit with a compound symmetry
covariance structure instead of calculating the statistics with the \strong{H} and
\strong{E} matrices.}
}
\value{
\tabular{ll}{
\code{EB} \tab A sum of squares cross products (SSCP) error matrix for
between-subject effects. \cr
\tab \cr
\code{E} \tab A sum of squares cross products (SSCP) error matrix for
within-subject effects. \cr
\tab \cr
\code{SSE} \tab The error sum of squares for within-subject effects calculated
as \eqn{SS_E = trace(}\strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1})}. \cr
\tab \cr
\code{spher} \tab A \code{data.frame} containing Mauchly's test for sphericity
results. \cr
\tab \cr
\code{ggEpsilon} \tab The Greenhouse-Geisser epsilon value. \cr
\tab \cr
\code{hflEpsilon} \tab The Huynh-Feldt-Lecoutre epsilon value. \cr
\tab \cr
\code{V} \tab The \strong{V} matrix from the marginal or mixed-effects model. \cr
\tab \cr
\code{M} \tab A sum-to-zero contrast matrix. \cr
\tab \cr
\code{n} \tab The sample size. \cr
\tab \cr
}
}
\description{
This function will calculate univariate tests for repeated measures ANOVA
(\code{mlm}), marginal (\code{\link[nlme]{gls}}) and mixed-effects
(\code{\link[nlme]{lme}}) models.  This includes Mauchly's test of sphericity
for within-subject effects and epsilon corrections for a departure from
sphericity.
}
\details{
\strong{H and E matrices}

In RM ANOVA, there are within-subject effects and (optionally) between-subject
effects.  An ANOVA can be done on a transformed dependent variable denoted as

\deqn{\Sigma(Y_i) / \sqrt r  for i = 1, \dots, r}

where the numerator is the sum of the repeated measures responses for each
subject and the denominator is the square root of the number of repeated
measurements (\eqn{r}). The between-subject effects are the independent variables in the ANOVA to
obtain the between-subject effects tests.

For within-subject effects, the SSCP error matrix is

\strong{E = M'} \eqn{(}\strong{Y'Y}\eqn{-} \strong{\eqn{\hat{B}}'} \eqn{(}\strong{X'X}\eqn{)}\strong{\eqn{\hat{B}}} \eqn{)}\strong{M}

and the SSCP hypothesis matrix is

\strong{H} \eqn{=} \strong{M'} \eqn{(}\strong{L\eqn{\hat{B}}} \eqn{)}'\eqn{(}\strong{L}\eqn{(}\strong{X'X}\eqn{) ^ {-}}\strong{L'} \eqn{) ^ {-1}}\strong{L\eqn{\hat{B}}M}

where \strong{M} is a sum-to-zero contrast matrix, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, and \strong{X} is the design matrix of the
independent variables.

For marginal and mixed effects models, the \strong{E} matrix is calculated from
the marginal variance-covariance matrix of the random/repeated effects
where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is
the corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix) as

\strong{E =} \eqn{n}\strong{M'VM}

and the \strong{H} matrix is calculated with the same formula as above.  The \strong{\eqn{\hat{B}}}
matrix is calculated by switching the reference levels of the within-subject
effect, refitting the models, and extracting the model coefficients.  The
contrast coefficients matrix (\strong{L}) is calculated from the \strong{LU}
decomposition of the crossproducts matrix (\strong{X'X}).

The F-statistics for the within-subject terms can be
calculated as \deqn{F = (SS_H / df_{Num}) / (SS_E / df_{Den})}

where \eqn{SS_E = trace(}\strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1})} is the error sum of
squares, the sum of squares for the hypothesis being tested is
\eqn{SS_H = trace(}\strong{H}\eqn{(}\strong{M'M}\eqn{) ^ {-1})}, and the numerator and
denominator degrees of freedom are denoted as \eqn{df_{Num}} and
\eqn{df_{Den}}, respectively.

\strong{Univariate Approach}

The marginal or mixed-effects model is refit with a compound symmetry
covariance structure for the univariate tests with the \code{refit = TRUE}
argument.

A Wald-type quadratic form is used to test each of the model terms.  The
Wald-type quadratic form for the F-tests is

\eqn{F = (}\strong{\eqn{\hat{\beta}}'L}\eqn{(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{) ^ {-1}}\strong{L'\eqn{\hat{\beta}}} \eqn{) / rank(}\strong{L}\eqn{)}

where \strong{\eqn{\hat{\beta}}} is a vector of fixed effects coefficients, \strong{L} is
a contrast coefficients matrix, \strong{X} is the fixed effects design matrix, and
\strong{V} is the marginal variance-covariance matrix
where \strong{V = ZGZ' + R} (fit with restricted maximum likelihood).  \strong{Z} is
the random effects design matrix, \strong{G} is the corresponding (co)variances of
the random effects matrix, and \strong{R} represents the (co)variances of the
repeated effects matrix.

The F-statistics for the terms in the \strong{H} and \strong{E} matrices formula above
can be calculated as \deqn{F = (SS_H / df_{Num}) / (SS_E / df_{Den})}
With this information, we can backsolve for the pseudo-sum of squares for the
hypothesis matrices in marginal and mixed-effects models as
\deqn{SS_H = (F(df_{Num})(SS_E)) / df_{Den}} to reproduce this result in the
univariate tests table.

\strong{Epsilon Corrections and Mauchly's Sphericity Test}

The within-subject SSCP \strong{E} matrix is used to calculate the
Greenhouse-Geisser and Huynh-Feldt-Lecoutre epsilons for univariate
adjustments for violations of sphericity as well as the sphericity test
itself.

The Greenhouse-Geisser epsilon is calculated as

\eqn{\epsilon_{GG} = ((\Sigma\lambda_i of} \strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1}) ^ 2) / (r - 1)(\Sigma\lambda_i ^ 2 of} \strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1})}

The Huynh-Feldt-Lecoutre epsilon is calculated as

\eqn{\epsilon_{HFL} = ((v + 1)(r - 1)\epsilon_{GG} - 2) / ((r - 1)(v - (r - 1)\epsilon_{GG}))}

Mauchly's test for sphericity is calculated for within-subject effects as

\eqn{W = (\Pi\lambda_i of} \strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1}) / (((\Sigma\lambda_i of} \strong{E}\eqn{(}\strong{M'M}\eqn{) ^ {-1}) / (r - 1)) ^ (r - 1))}

Let \eqn{r} be equal to the number of repeated measurements, \eqn{v}
be the residual degrees of freedom (between-subject design matrix), and the correction factor equal
\deqn{\rho = 1 - (2(r - 1) ^ 2 + (r - 1) + 2) / (6(r - 1)v)}  The \eqn{\chi ^ 2} statistic is
equal to the product \eqn{-\rho v log(W)} with degrees of freedom
\eqn{(r(r - 1) / 2) - 1}.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
uniTest(glsMod)

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
uniTest(lmeMod)

# Multivariate linear model
manovaMod <-
  manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
         data = orthodontWide)
uniTest(manovaMod)

}
\references{
{\url{https://www.lesahoffman.com/PSYC943/mv12psyc943_lecture13.pdf}}

{\url{https://go.documentation.sas.com/doc/en/pgmsascdc/9.4_3.3/statug/statug_glm_details46.htm}}

{\url{https://support.sas.com/rnd/app/stat/papers/mixedglm.pdf}}

{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}
}
\seealso{
{\code{\link[mmemmuris]{Hmatrix}}, \code{\link[mmemmuris]{coefs}}, \code{\link[mmemmuris]{Vmatrix}}, \code{\link[mmemmuris]{Ematrix}}, \code{\link[mmemmuris]{waldF}}, \code{\link[mmemmuris]{sphericityTests}}, \code{\link[mmemmuris]{epsilon}}, \code{\link[mmemmuris]{ddfResidual}}, \code{\link[mmemmuris]{ddfBW}}}
}
