% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{Vmatrix}
\alias{Vmatrix}
\title{Marginal Variance-Covariance Matrix for Random/Repeated Effects}
\usage{
Vmatrix(fMod, individual = NULL, full = FALSE, within = "Time")
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm}.}

\item{individual}{The subject whose marginal variance-covariance matrix
should be used, where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design
matrix, \strong{G} is the corresponding (co)variances of the random effects matrix,
and \strong{R} represents the (co)variances of the repeated effects matrix).  The
default is the first subject in the factor.}

\item{full}{If \code{TRUE}, print the full block diagonal matrix with all
experimental units.}

\item{within}{The name that should be used for the within-subject effect.
The default is "Time".}
}
\description{
This function calculates the marginal variance-covariance matrix of the
random/repeated effects for a marginal model (\code{gls}), mixed-effects model
(\code{lme}), and multivariate linear model (\code{mlm}).
}
\details{
The marginal variance-covariance matrix of the random/repeated effects is
denoted as

\strong{V = ZGZ' + R}

where \strong{Z} is the random effects design matrix, \strong{G} is the corresponding
(co)variances of the random effects matrix, and \strong{R} represents the
(co)variances of the repeated effects matrix.
}
\examples{
# Marginal model
# Unstructured covariance
glsModUN <- gls(Distance ~ Sex * Age,
                correlation =
                  corSymm(form = ~ as.numeric(Age) | Subject),
                weights = varIdent(form = ~ 1 | Age),
                na.action = na.omit,
                data = orthodontLong, method = "ML")

# Marginal variance-covariance matrix of the random/repeated effects for
# SSCP E matrix
Vmatrix(glsModUN)$V

# Fit with REML for marginal variance-covariance matrix of the
# random/repeated effects for compound symmetry covariance
glsModCS <- gls(Distance ~ Sex * Age,
                correlation =
                  corCompSymm(form = ~ 1 | Subject),
                na.action = na.omit,
                data = orthodontLong)
Vmatrix(glsModCS)

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)
Vmatrix(manovaMod)[c("unV", "csV")]
}
\references{
{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.3/statug/statug_mixed_details01.htm#statug_mixed009841}}
}
