% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{waldF}
\alias{waldF}
\title{Wald-Type F-Test with a Contrast Coefficients Matrix}
\usage{
waldF(fMod, L)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}} or \code{\link[nlme]{lme}}.}

\item{L}{A contrast coefficients matrix.}
}
\value{
\tabular{ll}{
\code{F} \tab A vector of Wald-type F-statistics. \cr
\tab \cr
\code{NumDF} \tab A vector of numerator degrees of freedom. \cr
\tab \cr
\code{X2} \tab A vector of Wald chi-square statistics.
\cr \tab \cr
}
}
\description{
This function calculates Wald-type F-statistics and Wald chi-square
statistics for a model.
}
\details{
For marginal and mixed-effects models, we start with a Wald-type quadratic
form

\eqn{F = (}\strong{\eqn{\hat{\beta}}'L}\eqn{(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{) ^ {-1}}\strong{L'\eqn{\hat{\beta}}} \eqn{) / rank(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{)}

where \strong{\eqn{\hat{\beta}}} is a vector of fixed effects coefficients, \strong{X} is the fixed effects design matrix, and
\strong{V} is the marginal variance-covariance matrix
where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is
the corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix).  The contrast coefficients matrix (\strong{L}) is calculated from the \strong{LU}
decomposition of the crossproducts matrix (\strong{X'X}).

It follows that the Wald-\eqn{\chi ^ 2} statistic is the numerator degrees of
freedom (\eqn{rank(}\strong{L'} \eqn{(}\strong{X'V}\eqn{ ^ {-1}}\strong{X}\eqn{) ^ {-}}\strong{L}\eqn{)})
multiplied by the F-statistic.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)

# Wald chi-square statistic
L <- mmemmuris:::Lmatrix.ulm(glsMod)
L$Sex <- rbind(L$Sex, L$`Sex:Age`)
waldF(glsMod, L)

# Wald chi-square statistic
hlTrace(glsMod, approximation = "W")

# Wald chi-square statistic
Ftests <- anova(glsMod)[c("Age", "Sex:Age"), ]
Ftests$Chisq <- Ftests$numDF * Ftests$`F-value`
Ftests$`Pr(>Chisq)` <- 1 - pchisq(Ftests$Chisq, Ftests$numDF)
Ftests

}
\seealso{
\code{\link[mmemmuris]{hlTrace}}
}
