% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmemmuris.R
\name{wilksLambda}
\alias{wilksLambda}
\title{Multivariate Tests: Wilks' Lambda}
\usage{
wilksLambda(
  fMod,
  individual = NULL,
  approximation = c("Rao", "Bartlett", "LR"),
  within = "Time",
  LR = TRUE,
  ss = NULL,
  tolerance = .Machine$double.eps
)
}
\arguments{
\item{fMod}{A model of class \code{\link[nlme]{gls}}, \code{\link[nlme]{lme}}, or \code{mlm}.}

\item{individual}{The subject whose marginal variance-covariance matrix
should be used, where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design
matrix, \strong{G} is the corresponding (co)variances of the random effects matrix,
and \strong{R} represents the (co)variances of the repeated effects matrix).  The
default is the first subject in the factor.}

\item{approximation}{Approximation used for Wilks' \eqn{\Lambda}.  Options are Rao F ("Rao"),
Bartlett-corrected likelihood ratio \eqn{\chi ^ 2} ("Bartlett"), and likelihood
ratio \eqn{\chi ^ 2} ("LR").}

\item{within}{The name that should be used for the within-subject effect.
The default is "Time".}

\item{LR}{Likelihood ratio tests are used to calculate the likelihood ratio
\eqn{\chi ^ 2} and Wilks' \eqn{\Lambda} statistics instead of the \strong{H} and
\strong{E} matrices for marginal and mixed-effects models.}

\item{ss}{The sample size of the dataset.  If \code{NULL}, refer to \code{\link[mmemmuris]{completeData}}
for calculation of the sample size.  Otherwise, use with care.}

\item{tolerance}{Threshold for detecting columns that are linearly dependent.
Use with care.}
}
\value{
This function will return two lists: a within-subject effects list
(\code{withinTests}) and a between-subject effects list (\code{betweenTests}) if
between-subject effects are in the model.
\code{withinTests} is a list containing: \tabular{ll}{
\code{wilks}  \tab A \code{data.frame} containing the Wilks' \eqn{\Lambda} statistics and
hypothesis tests for these statistics. \cr
\code{parmsWilks}  \tab A \code{data.frame} containing the values used to calculate
the Wilks' \eqn{\Lambda} statistics. \cr
\code{approximation}  \tab Approximation used for the Wilks' \eqn{\Lambda} statistics.  Options are Rao F ("Rao"),
Bartlett-corrected likelihood ratio \eqn{\chi ^ 2} ("Bartlett"), and likelihood
ratio \eqn{\chi ^ 2} ("LR"). \cr
\code{E}  \tab The sums of squares and crossproducts error matrix for within-subject effects. \cr
\code{H}  \tab The sums of squares and crossproducts hypothesis matrices for within-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\code{type}  \tab The type of \strong{L} matrix used.  Type 1 is sequential. \cr
\tab \cr
}
\code{betweenTests} is a list containing: \tabular{ll}{
\code{wilks}  \tab A \code{data.frame} containing the Wilks' \eqn{\Lambda} statistics and
hypothesis tests for these statistics. \cr
\code{parmsWilks}  \tab A \code{data.frame} containing the values used to calculate
the Wilks' \eqn{\Lambda} statistics. \cr
\code{approximation}  \tab Approximation used for the Wilks' \eqn{\Lambda} statistics.  Options are Rao F ("Rao"),
Bartlett-corrected likelihood ratio \eqn{\chi ^ 2} ("Bartlett"), and likelihood
ratio \eqn{\chi ^ 2} ("LR"). \cr
\code{EB}  \tab The sums of squares and crossproducts error matrix for between-subject effects. \cr
\code{HB}  \tab The sums of squares and crossproducts hypothesis matrices for between-subject effects. \cr
\code{L}  \tab The contrast coefficients matrices. \cr
\code{type}  \tab The type of \strong{L} matrix used.  Type 1 is sequential. \cr
\code{Vmatrix}  \tab The marginal variance-covariance matrix of the random/repeated effects. \cr
\tab \cr
}
}
\description{
This function will calculate Wilks' \eqn{\Lambda} statistics from marginal models,
mixed-effects models, and multivariate linear models.
}
\details{
\strong{H and E matrices}

The Wilks' lambda statistic for RM ANOVA is calculated as

\eqn{\Lambda = (det(}\strong{E}\eqn{)) / (det(}\strong{H}\eqn{+}\strong{E}\eqn{))}

where the SSCP error matrix is

\strong{E = M'} \eqn{(}\strong{Y'Y}\eqn{-} \strong{\eqn{\hat{B}}'} \eqn{(}\strong{X'X}\eqn{)}\strong{\eqn{\hat{B}}} \eqn{)}\strong{M}

and the SSCP hypothesis matrix is

\strong{H} \eqn{=} \strong{M'} \eqn{(}\strong{L\eqn{\hat{B}}} \eqn{)}'\eqn{(}\strong{L}\eqn{(}\strong{X'X}\eqn{) ^ {-}}\strong{L'} \eqn{) ^ {-1}}\strong{L\eqn{\hat{B}}M}

where \strong{M} is an identity matrix for between-subject effects and a
sum-to-zero contrast matrix for within-subject effects, \strong{Y} is a combined
matrix of observations, \strong{\eqn{\hat{B}}} is a matrix of combined coefficients from the
ordinary least squares models, and \strong{X} is the design matrix of the
independent variables.

For marginal and mixed effects models, the \strong{E} matrix is calculated from
the marginal variance-covariance matrix of the random/repeated effects
(\strong{V})

\strong{E =} \eqn{n}\strong{M'VM}

where \strong{V = ZGZ' + R} (\strong{Z} is the random effects design matrix, \strong{G} is
the corresponding (co)variances of the random effects matrix, and \strong{R}
represents the (co)variances of the repeated effects matrix).  The \strong{H}
matrix is calculated with the same formula as above.  The \strong{\eqn{\hat{B}}} matrix is
calculated by switching the reference levels of the within-subject effect,
refitting the models, and extracting the model coefficients.

\strong{Likelihood Ratio}

The Wilks' lambda statistic for a marginal or mixed-effects model is
essentially a likelihood ratio test with the reduced model excluding the
term being tested.  The log-likelihood function for the full model is

\eqn{logLik(M_F) = (-1 / 2)log(det(}\strong{V_{F}} \eqn{)) - ((1 / 2)}\strong{r'V_{F}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

where \strong{V_{F}} is the marginal variance-covariance matrix of the random/repeated
effects, \strong{r} is the vector of model residuals, and \eqn{N} is the number of
observations of data in the long format.

The contrast coefficients matrix (\strong{L}) is calculated from the \strong{LU}
decomposition of the crossproducts matrix (\strong{X_{F}'X_{F}}).  To
project the design matrix of the reduced model (\strong{X_{R}}) onto the
column space of the design matrix of the full model, the design matrix of the
full model needs to be multiplied by the orthogonal complement of \strong{L'}.
It follows that the log-likelihood function for the reduced model is

\eqn{logLik(M_R) = (-1 / 2)log(det(}\strong{V_{R}} \eqn{)) - ((1 / 2)}\strong{r'V_{R}} \eqn{^ {-1}}\strong{r} \eqn{) - (N / 2)log(2\pi)}

where \strong{V_{R}} is the marginal variance-covariance matrix of the random/repeated
effects, \strong{r} is the vector of model residuals, and \eqn{N} is the number of
observations of data in the long format.

Wilks' \eqn{\Lambda} is denoted as
\deqn{\Lambda = exp{(-[-2logLik(M_R) - \{-2logLik(M_F)\}]) / n}}
where \eqn{n} is the number of subjects.

The likelihood ratio (LR) statistic approximately follows a \eqn{\chi ^ 2}
distribution \deqn{LR = -2logLik(M_R) - \{-2logLik(M_F)\}} with
\eqn{df_{M_R} - df_{M_F}} degrees of freedom.

\strong{Rao}

Rao's approximation approximately follows an F-distribution
\deqn{F_\Lambda = [(1 - \Lambda ^ {1 / t}) / (\Lambda ^ {1 / t})][(rt - 2u) / (pq)]}
where \eqn{p = rank(}\strong{H}\eqn{+}\strong{E}\eqn{)}, \eqn{q = rank(}\strong{L}\eqn{(}\strong{X'X}\eqn{)^{-1}}\strong{L'} \eqn{)}, \eqn{v}
is the residual degrees of freedom, \eqn{r = v - (p - q + 1) / 2},
\eqn{u = (pq - 2) / 4}, and

if \eqn{p ^ 2 + q ^ 2 - 5 > 0}, \deqn{t = \sqrt{(p ^ 2q ^ 2 - 4) / (p ^ 2 + q ^ 2 - 5)}}
else \deqn{t = 1}

with \eqn{pq} numerator degrees of freedom and \eqn{rt - 2u} denominator
degrees of freedom.

If \eqn{s = min(p, q) \le 2}, the F-statistic is exact.

\strong{Bartlett-Corrected Likelihood Ratio}

The Bartlett-corrected LR statistic approximately follows a \eqn{\chi ^ 2}
distribution \deqn{LR_B = - (((p - q + 1) / 2) - v)ln(\Lambda)} with
\eqn{pq} degrees of freedom.
}
\examples{
# Marginal model
# Unstructured covariance
glsMod <- gls(Distance ~ Sex * Age,
              correlation =
                corSymm(form = ~ as.numeric(Age) | Subject),
              weights = varIdent(form = ~ 1 | Age),
              na.action = na.omit,
              data = orthodontLong)
# Rao F statistic
wilksLambda(glsMod)
# Bartlett-corrected likelihood ratio chi-square statistic
wilksLambda(glsMod, approximation = "B")
# Likelihood ratio chi-square statistic
wilksLambda(glsMod, approximation = "LR")

# Linear mixed-effects model
# Unstructured covariance
lmeMod <- lme(Distance ~ Sex * Age,
              random = ~ Age | Subject,
              na.action = na.omit,
              data = orthodontLong)
# Rao F statistic
wilksLambda(lmeMod)
# Bartlett-corrected likelihood ratio chi-square statistic
wilksLambda(lmeMod, approximation = "B")
# Likelihood ratio chi-square statistic
wilksLambda(lmeMod, approximation = "LR")

# Multivariate linear model
manovaMod <- manova(cbind(Distance8, Distance10, Distance12, Distance14) ~ Sex,
                    data = orthodontWide)
# Rao F statistic
wilksLambda(manovaMod)
# Bartlett-corrected likelihood ratio chi-square statistic
wilksLambda(manovaMod, approximation = "B")
# Likelihood ratio chi-square statistic
wilksLambda(manovaMod, approximation = "LR")

}
\references{
{\url{https://support.sas.com/resources/papers/proceedings/proceedings/sugi23/Stats/p229.pdf}}

{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.4/statug/statug_introreg_sect038.htm}}

{\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.3/statug/statug_mixed_details01.htm#statug.mixed.mixedmmtgandr}}

{Rao, C. R. (1973). Linear Statistical Inference and Its Applications. 2nd ed. New York: John Wiley & Sons.}
}
\seealso{
{\code{\link[mmemmuris]{Hmatrix}}, \code{\link[mmemmuris]{coefs}}, \code{\link[mmemmuris]{Vmatrix}}, \code{\link[mmemmuris]{Ematrix}}, \code{\link[mmemmuris]{REMLtoML}}, \code{\link[mmemmuris]{LRT}}}
}
